worked on all files in src folder
