+++ src/components/AboutDeiBadgingCom/About.js
@@ -9,14 +9,7 @@ const AboutNew = () => {
   const { isDesktop } = useDesktop();
 
   return (
-    <div className="about-dei">
-      {/* <div className="header">
-        <Header />
-        <div className="container jumbotron__container">
-          <img src={curlyBraces} alt="badging-logo" />
-          <h1>About CHAOSS DEI Project Badging</h1>
-        </div>
-      </div> */}
+    <div className="about-dei">    
       <main className="container">
         <div className="main__container">
           {isDesktop && (
@@ -36,7 +29,6 @@ const AboutNew = () => {
           )}
         </div>
       </main>
-      {/* <Footer /> */}
     </div>
   );
 };
