+++ src/components/Faq/faq.scss
@@ -30,7 +30,6 @@
     p:nth-child(1) {
       display: flex;
       justify-content: space-between;
-      // background-color: red;
       width: inherit;
       height: inherit;
       align-items: center;
