+++ src/components/Footer/footer.scss
@@ -1,7 +1,5 @@
 .footer-container {
-    // position: absolute;
-    // bottom: 0;
-    // width: 100%;
+    
     gap: 0 !important;
     
     .social {
@@ -20,7 +18,7 @@
     }
     }
 
-    .copyright {
+    .copy-right {
         padding: 15px 10px;
         background-color: #BDFFF5;
         color: #000;
