+++ src/components/Header/Card2.scss
@@ -62,11 +62,7 @@
         font-weight: bold;
       }
     }
-    // .hidden_title p {
-    //   padding-left: 25px;
-    //   padding-top: 10px;
-    //   // color: #11be34;
-    // }
+   
     .card-title {
       position: absolute;
       background-color: black;
@@ -87,7 +83,6 @@
     }
     .copy {
       border: 1px solid #454c53;
-      // display: none;
       border-bottom: 0;
       margin: 60px 25px 0 25px;
       height: 120px;
