+++ src/components/Header/Header.js
@@ -21,7 +21,7 @@ const Nav = () => {
           <li><a href="/project-badging" className={pathname.includes('/project-badging') ? 'activeNav': ''}>Project Badging</a></li>
           <li><a href="/faq" className={pathname == '/faq'? 'activeNav': ''}>FAQ</a></li>
         </ul>
-        <img className='mobileNav' onClick={() => setToggleNav(!toggleNav)} src={mobileMenuIcon} alt="logo" />
+        <img className='mobile-nav' onClick={() => setToggleNav(!toggleNav)} src={mobileMenuIcon} alt="logo" />
         {
         pathname && (pathname.includes('/event-badging') || pathname.includes('/project-badging')) ? (
           <a href={pathname.includes('/event-badging') ? 'https://chaoss.community/diversity-and-inclusion-badging/' : '/badge'} className='apply-btn'>Apply for Badge</a>
@@ -31,7 +31,7 @@ const Nav = () => {
       }
 
      </header>
-     <div className="mobileNavContainer" style={{display: toggleNav ==true ? '' : 'none'}}>
+     <div className="mobile-nav-container" style={{display: toggleNav ==true ? '' : 'none'}}>
           <ul className='mobile-list'>
             <li><a href="/" className={pathname == '/'? 'activeNav': ''}>Home</a></li>
             <li><a href="/event-badging" className={pathname == '/event-badging'? 'activeNav': ''}>Event Badging</a></li>
