+++ src/components/Header/header.scss
@@ -9,14 +9,14 @@ header {
 	position: fixed;
 	background-color: rgba(255, 255, 255, 0); /* Semi-transparent background for the glass effect */
 	backdrop-filter: blur(10px); /* Adjust the blur amount as needed */
-	// box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+
 	
 	top: 0;
 	right: 0;
 	left: 0;
 	width: 100%;
 	
-  //   width: 100%;
+  
   
 	ul {
 	  display: flex;
@@ -52,7 +52,7 @@ header {
       }
 	}
 
-	.mobileNav {
+	.mobile-nav {
 		background-color: #000;
 		padding: 12px;
 		border-radius: 50px;
@@ -62,7 +62,7 @@ header {
 
   }
 
-  .mobileNavContainer {
+  .mobile-nav-container {
 	z-index: 99;
 	width: 100%;
 	height: 400px;
@@ -73,8 +73,7 @@ header {
 	border-bottom-left-radius: 10px;
 	border-bottom-right-radius: 10px;
 	
-    // background: radial-gradient(50% 50% at 50% 50%, rgba(98, 203, 251, 0.30) 0%, rgba(255, 255, 255, 0.00) 62.81%);
-    
+  
 	
 	.mobile-list {
 		margin-top: 3rem;
