+++ src/components/SelectProject/search-bar.scss
