+++ src/components/index.js
@@ -2,16 +2,16 @@ import Header from "./Header/Header";
 import Footer from "./Footer/Footer";
 import Faq from "./Faq/Faq";
 import CardComponent from "./HowItWorks/Card";
-import CardCarousel from "./Carousel/CardCarousel";
-import SearchBar from "./SelectProject/SearchBar";
-import Sidebar from "./AboutDeiBadging/Sidebar";
+import CardCarousel from "./Carousel/Card_Carousel";
+import SearchBar from "./SelectProject/Search_Bar";
+import Sidebar from "./AboutDeiBadging/Side_bar";
 import Jumbotron from "./Jumbotron/Jumbotron";
-import AboutDeiMobile from "./AboutDeiBadging/AboutDeiMobile";
-import WhatIsDeiBadging from "./AboutDeiBadging/WhatIsDeiBadging";
-import ApplicationProcess from "./AboutDeiBadging/ApplicationProcess";
-import DeiFile from "./AboutDeiBadging/DeiFile";
-import DeiBadge from "./AboutDeiBadging/DeiBadge";
-import SelectedProjects from "./SelectProject/SelectedProjects";
+import AboutDeiMobile from "./AboutDeiBadging/About_Dei_Mobile";
+import WhatIsDeiBadging from "./AboutDeiBadging/What_Is_Dei_Badging";
+import ApplicationProcess from "./AboutDeiBadging/Application_Process";
+import DeiFile from "./AboutDeiBadging/Dei_File";
+import DeiBadge from "./AboutDeiBadging/Dei_Badge";
+import SelectedProjects from "./SelectProject/Selected_Projects";
 import Layout from "./Layout/Layout";
 import Spinner from "./Loader/Spinner";
 import Loader from "./Loader/Loader";
