+++ src/pages/EventBadging/event-badging.scss
