+++ src/pages/GetStartedBadging/get-started-badging.scss
