+++ src/pages/NewBadgingRequirements/latest-badging.scss
