+++ src/pages/index.js
@@ -1,18 +1,16 @@
 import Home from "./Home/Home";
-// import DeiBadge from './DeiBadge/DeiBadge';
-
 import Projects from "./Projects/Projects";
 
-import GetStartedBadging from "./GetStartedBadging/GetStartedBadging";
+import GetStartedBadging from "./GetStartedBadging/Get_Started_Badging";
 
 import About from "./AboutDeiBadging/About";
 
-import SelectProjectRepo from "./SelectProjectRepo/SelectProjectRepo";
+import SelectProjectRepo from "./SelectProjectRepo/Select_Project_Repo";
 
 import ErrorPage from "./Home/ErrorPage/ErrorPage";
 
-import SuccessfullyBadged from "./ProjectBadgingSuccess/ProjectBadgingSuccess";
-import ComingSoon from "./ComingSoon/ComingSoon";
+import SuccessfullyBadged from "./ProjectBadgingSuccess/Project_Badging_Success";
+import ComingSoon from "./ComingSoon/Coming_Soon";
 
 export {
   Home,
